+++ _temp_feature_discussion.txt
@@ -13,7 +13,7 @@ discussion -----------------------------------------
 - display: showpath='all'
 - display: graphic defaults to config
 
-- testing code coverage -> 100
+- testing code coverage -> 99.9% -> 100% how to test plt.plot()
 
 - code: apply .reshape((-1,3))
 - code: numpy reshape and other operations creating copies
@@ -57,6 +57,8 @@ code: .misc package
     - Dipole
     - FEM interpolation
     - Virtual field (just a function)
+    - custom source for given field
+
 
 - potential performance increase:
     Can we speed up the computation by combining similar
