+++ tests/test_BaseGeo.py
@@ -1,11 +1,12 @@
+import os
+import pickle
 import numpy as np
 from scipy.spatial.transform import Rotation as R
 from magpylib._lib.obj_classes.class_BaseGeo import BaseGeo
 
 
-def test_BaseGeo():
-    """
-    fundamental simple test of setter, getter, .move and .rotate
+def test_BaseGeo_basics():
+    """ fundamental usage test
     """
 
     ptest = np.array([[0,0,0], [1,2,3], [0,0,0], [0,0,0],
@@ -104,3 +105,51 @@ def test_BaseGeo_neg_steps():
     bg.move_by((1,2,3),steps=3)
     bg.move_by((1,2,3),steps=-5)
     assert len(bg.pos)==4, 'bad negative steps'
+
+
+def test_path_BaseGeo():
+    """test path functionality
+    """
+    # pylint: disable=protected-access
+    poss = []
+    rots = []
+
+    bgeo = BaseGeo((0,0,0), None)
+    poss += [bgeo.pos]
+    rots += [bgeo.rot.as_rotvec()]
+    poss += [bgeo._pos]
+    rots += [bgeo._rot.as_rotvec()]
+
+    bgeo = BaseGeo([(0,0,0),(1,1,1)], None)
+    poss += [bgeo.pos]
+    rots += [bgeo.rot.as_rotvec()]
+    poss += [bgeo._pos]
+    rots += [bgeo._rot.as_rotvec()]
+
+    bgeo = BaseGeo((0,0,0), R.from_rotvec((.1,.2,.3)))
+    poss += [bgeo.pos]
+    rots += [bgeo.rot.as_rotvec()]
+
+    bgeo.pos = [(2,2,2),(3,3,3)]
+    bgeo.rot = R.from_rotvec([(.1,.2,.3),(.2,.3,.4)])
+    poss += [bgeo.pos]
+    rots += [bgeo.rot.as_rotvec()]
+
+    bgeo = BaseGeo((0,0,0), None)
+    bgeo.move_by((3,0,0))
+    bgeo.move_by((3,0,0), steps=3)
+    bgeo.rotate_from_angax(330,(0,0,1),anchor=0,steps=5)
+    bgeo.move_by((0,0,10),steps=-7)
+    poss += [bgeo.pos]
+    rots += [bgeo.rot.as_rotvec()]
+
+    x = np.array([])
+    for p,r in zip(poss,rots):
+        x = np.r_[x,p.flatten(),r.flatten()]
+
+    #pickle.dump(x,open('testdata_path_BaseGeo.p', 'wb'))
+
+    x_test = pickle.load(open(
+        os.path.abspath('tests/testdata/testdata_path_BaseGeo.p'), 'rb'))
+
+    assert np.allclose(x,x_test), 'path problem'
