+++ tests/test_field_BH_box.py
@@ -44,7 +44,8 @@ def test_field_BH_box():
     """ test box field
     """
     mag3.Config.EDGESIZE=1e-14
-    mag, dims, poss, B = pickle.load(open(os.path.abspath('tests/testdata/testdata_field_BH_box.p') ,'rb'))
+    mag, dims, poss, B = pickle.load(open(
+        os.path.abspath('tests/testdata/testdata_field_BH_box.p') ,'rb'))
     Btest = []
     for dim,pos in zip(dims,poss):
         Btest += [field_BH_box(True, mag, dim, pos)]
